I saw the tests complain with node6.
Looks like async-await is needed, and that was probably not available then.
I am not sure how the project wants to ensure the node 6.9 compat, as stated
in the package.json, so I just jumped to latest node to be able to run the
project. Let's see if a PR review sheds some light.